fix error on export csv campaign summary report

+++ app/controllers/api/v1/CampaignReportAPIController.php
@@ -532,10 +532,10 @@ class CampaignReportAPIController extends ControllerAPI
             $totalPageViews = (int) isset($total[0]->page_views)?$total[0]->page_views:0;
 
             // Get total estimate
-            $totalEstimated = (int) isset($total[0]->estimated_total)?$total[0]->estimated_total:0;
+            $totalEstimated = isset($total[0]->estimated_total)?$total[0]->estimated_total:0;
 
             // Get total spending
-            $totalSpending = (int) isset($total[0]->spending)?$total[0]->spending:0;
+            $totalSpending = isset($total[0]->spending)?$total[0]->spending:0;
 
             $_campaign->select('campaign_id');
